fix(cli): don't let mail passthrough swallow autostart/heartbeat global flags
The first cut added `mail` to RAW_SUBCOMMANDS and moved the "collect every
trailing token verbatim" guard to the top of the parse loop. That fixed
`mail connect --email …` but regressed the two sibling raw subcommands:
`autostart install --port/--channels/--imessage` and `heartbeat run --model`
read those *recognized global* flags off the parsed fields, and the top-of-loop
guard now shovelled them into subargs before the parser saw them — so
`autostart install --port 8080 --channels imessage` silently installed on the
default 5757 with no channels.

mail (re-parses all its own args) and autostart/heartbeat (want a few global
flags parsed) have opposite needs, so split them:
- PASSTHROUGH_SUBCOMMANDS = {mail} drives the top-of-loop verbatim collection.
- RAW_SUBCOMMANDS = {heartbeat, autostart} keeps the narrow behavior from main —
  only *unrecognized* trailing flags are collected, in the --flag branch.

Also extract the pure parser to src/cli-args.ts (cli.ts runs main() on import,
so the exported parseArgs wasn't actually importable for tests) and add
src/cli-args.test.ts covering mail passthrough, autostart/heartbeat global
flags, global-before-subcommand, unknown-flag throw, and the value-vs-subcommand
guard.
